backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.77",
    "celery>=5.6.3",
    "colorlog>=6.10.1",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-core>=1.2.22",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "numpy>=2.4.3",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.4.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
